+++ docs/projects/robot/code/rpi/robot/__init__.py
@@ -1,5 +1,8 @@
-from .param import const, PARAM_RESERVED, PARAM, PARAM_FS, PARAM_PWM, \
-        STATE_K, STATE_PITCH, STATE_CPT1, STATE_CPT2, \
-        STATE_DUTY1, STATE_DUTY2, STATE_DT1, STATE_DT2
-from .comm import Comm
-from .remote import Remote
+# fix wiring issue
+import os
+if os.getenv('BALENA_DEVICE_ARCH') == 'aarch64':
+    from gpiozero import Button as Pin
+    Pin(14, pull_up=False)
+    
+from . abstract_robot import AbstractRobot
+from . remote import Remote