+++ docs/projects/robot/code/rpi/robot/remote.py
@@ -1,25 +1,23 @@
 from .ble_uart import BLE_UART
 from struct import pack, unpack
-import asyncio, time
+import asyncio, time, os
+
+# fix robot wiring issue (if there is one?)
+if os.getenv('BALENA_DEVICE_ARCH') == 'aarch64':
+    try:
+        from gpiozero import Button as Pin
+        Pin(14, pull_up=False)
+    except:
+        pass
 
 
 class Remote:
     
-    def __init__(self):
+    def __init__(self, event_handler=None):
         self.stop = False
+        self._event_handler = event_handler
         
-    async def handle(self, dt: float, code: str, value: float):
-        """Handle message received from BLE. 
-        Overload in derived class.
-        @param dt: time [s] since last message
-        @param code: character code of message
-        @param value: value of message
-        """
-        if code != 'h':
-            # if dt < 0.01: print(f"***** QUICK READ:  {1000*dt:12.8} [ms]")
-            print(f"RECV {code} = {value:8.3f}   dt = {1000*dt:12.4} [ms]")
-            
-    async def shutdown(self, dt: float, code: str):
+    def shutdown(self, code: str):
         """Called if remote disconnects - turned off (code='q') or connection lost ('d').
         Overload in derived class.
         After this call, self.run returns.
@@ -43,22 +41,19 @@ class Remote:
         self.stop = True
 
     async def _recv(self):
-        last_rx = time.monotonic()
         while not self.stop:
             try:
                 data = await asyncio.wait_for(self.uart.read(), timeout=2)
-                dt = time.monotonic() - last_rx
-                last_rx = time.monotonic()
                 if data == None: 
-                    code = 'd'      
+                    code, value = ('d', 0)      
                 else:
                     code, value = unpack('>Bf', data)
                     code = chr(code)
+                if self._event_handler:
+                    await self._event_handler(code, value)
                 if code == 'q' or code == 'd':
-                    await self.shutdown(dt, code)
+                    self.shutdown(code)
                     break
-                else:
-                    await self.handle(dt, code, value)
             except asyncio.TimeoutError:
                 await self.timeout()
         self.stop = True
@@ -73,19 +68,22 @@ class Remote:
         except Exception as e:
             print(f"***** remote.rgb: {e}")
             
-            
-    async def run(self, peripheral_name='iot49-robot'):
-        """Sample main"""
-        async with BLE_UART(peripheral_name) as uart:
-            self.uart = uart
-            await uart.connect()
-            asyncio.create_task(self._recv())
-            blue = 0
-            while not self.stop:
-                await asyncio.sleep(0.5)
-            print("so long ...")
-            try:
-                await uart.disconnect()
-            except Exception as e:
-                print("# disconnect barfed", type(e))
-                        
+    async def __aenter__(self):
+        self.uart = BLE_UART('iot49-robot')
+        await self.uart.connect()
+        self._recv_task = asyncio.create_task(self._recv(), name="remote._recv")
+        self._recv_task.add_done_callback(self._done_callback)
+        return self
+
+    async def __aexit__(self, *args):
+        self._recv_task.cancel()
+        await self.uart.disconnect()
+        return self
+
+    def _done_callback(self, task: asyncio.Task) -> None:
+        try:
+            task.result()
+        except asyncio.CancelledError:
+            pass
+        except Exception as e:
+            print(f"***** Task {task.get_name()}: {repr(e)}")