+++ docs/projects/robot/code/stm32/lib/no_control.py
@@ -1,0 +1,7 @@
+from robot import Controller
+
+
+class Control(Controller):
+
+    def update(self):
+        pass