+++ docs/projects/robot/code/stm32/lib/robot/__init__.py
@@ -1,0 +1,4 @@
+from . comm import Comm as Robot
+from . controller import Controller
+from . pid import *
+from . state import *