+++ docs/projects/robot/code/stm32/lib/speed_control.py
@@ -1,18 +1,22 @@
-from controller import *
-from param import PARAM
-
-PARAM_DUTY1  = const(2)   # motor1 duty cycle setpoint
-PARAM_DUTY2  = const(3)   # motor2 duty cycle setpoint
+from robot import STATE, DUTY_CM, DUTY_DEAD, TACHO_CM, PID_TACH_CM
+from robot import Controller, PID
+from robot import SETPOINT, KP
 
 class Control(Controller):
     
-    @staticmethod
-    def fix_duty(duty):
-        if abs(duty) > 1:
-            duty = duty+8.5 if duty>0 else duty-8.5
-        return duty
-
+    def __init__(self, uart):
+        super().__init__(uart)
+        self.dz  = STATE[DUTY_DEAD]
+        self.pid = PID(memoryview(STATE)[PID_TACH_CM:])
+        
     def update(self):
-        # set motor duty cycle
-        self.state[STATE_DUTY1] = self.fix_duty(PARAM[PARAM_DUTY1])
-        self.state[STATE_DUTY2] = self.fix_duty(PARAM[PARAM_DUTY2])
+        duty = self.pid.update(STATE[TACHO_CM])
+        # STATE[DUTY_CM] = self.pid.config[SETPOINT]
+        STATE[DUTY_CM] = self.fix_duty(duty)
+
+    
+    def fix_duty(self, duty):
+        if abs(duty) > 0.5:
+            DZ = self.dz
+            return duty+DZ if duty>0 else duty-DZ
+        return 0